refactor: 자동차 경주 미션 default 이동 전략 별도 파일 분리
